Added new folder to Source: ThirdParty. Added SDL3, libpng, turbojpen/libjpeg and zlib libraries to ThirdParty. Swtiched from Windows API to SDL3 for window management. Implemented new SystemWindow class to replace the now removed IWindow class. Added to FileStream class static functions: readJpegImage() and readPngImage(). Those two functions implement the turbojpg and libpng libraries to read data from both png and jpeg files. Linked SDL window to DirectX API, window so the engine's rendering still works as intended and no changes where made to any graphic related modules to implement to switchto SDL.

+++ Source/Runtime/Direct3D11/Private/Resources/Primitives.h
@@ -115,7 +115,7 @@ inline Mesh createFlatCircle(int precision = 0)
   memcpy(pIndexData, indices.data(), sizeof(uint16) * indices.size());
 
   instance.vertexBuffer = VertexBuffer(pVerts, sizeof(SGFXVertex) * (int)verts.size());
-  instance.indexBuffer = IndexBuffer(pIndexData,  sizeof(uint16) * indices.size());
+  instance.indexBuffer = IndexBuffer(pIndexData,  sizeof(uint16) * static_cast<int>(indices.size()));
 
   return instance;
 }
